ChatServiceCommunication_Version-1.0.1
Changed Key for Private Chat,
The hope was that this would have made it so that the actual IP was used instead of 0.0.0.0 but that wasnt the case.
BUT this way it automatically combines the Ip and Port so that is ok too. But it will probably stil cause problems with more than 1 server, The Solution would probably be to get the Pod ip when the Gameserver was made, but since we use the Backup Game Lobby that isnt going to happen